requirements.txt
pylint
yapf
pytesseract
opencv-python
pillow
imutils
rope